fix the following bug
-	model_->bias = (int)ptr[0];
+	model_->bias = ptr[0];

+++ matlab/linear_model_matlab.c
@@ -145,7 +145,7 @@ const char *matlab_matrix_to_model(struct model *model_, const mxArray *matlab_s
 
 	// bias
 	ptr = mxGetPr(rhs[id]);
-	model_->bias = (int)ptr[0];
+	model_->bias = ptr[0];
 	id++;
 
 	if(model_->bias>=0)
